feat: restore dynamic routes and update static route to evaluate errors
