+++ pages/[...slug].tsx
@@ -1,46 +1,34 @@
-import { GetStaticPaths, GetStaticProps } from 'next';
+import { GetServerSideProps, GetServerSidePropsContext } from 'next';
+import { generateMdx, getDefaultTheme, getPageContent } from '@cmpsr/nextjs-contentful-renderer';
 import { Page } from '../src/Page';
-import { staticPaths } from '../src/staticPaths';
-import { getStaticPropsContent } from '../src/getStaticPropsContent';
 
-export const getStaticPaths: GetStaticPaths = async () => {
-  const paths = await staticPaths();
-  const filteredPaths = paths.filter((path) => path !== '/');
-
-  return {
-    paths: filteredPaths.map((slug) => ({
-      params: { slug: slug.split('/').slice(1) },
-    })),
-    fallback: 'blocking',
+export const getServerSideProps: GetServerSideProps<any> = async (context: GetServerSidePropsContext) => {
+  const notFound: { notFound: true } = {
+    notFound: true,
   };
-};
-
-const getSlug = (path: string | string[]): string => (Array.isArray(path) ? path.join('/') : path);
-
-export const getStaticProps: GetStaticProps = async ({ preview, params }) => {
   try {
-    return await getStaticPropsContent(getSlug(params.slug), preview || false);
-  } catch (err) {
-    if (isTooManyRequestsError(err)) {
-      return await getStaticProps({ preview, params });
-    } else {
-      console.error(err);
-      throw err;
+    const page = await getPageContent(context);
+    if (!page) {
+      return notFound;
     }
+    let theme = page.theme;
+    if (!theme) {
+      theme = await getDefaultTheme(context.query.preview !== undefined);
+    }
+
+    const content = await generateMdx(page.content, page.globalVariables);
+    return {
+      props: {
+        title: page.title,
+        content,
+        metaConfiguration: page.metaConfiguration || [],
+        theme: theme || null,
+      },
+    };
+  } catch (e) {
+    console.error('Dynamic route error', JSON.stringify(e));
+    return notFound;
   }
 };
+
 export default Page;
-
-const isTooManyRequestsError = (err: any): boolean => {
-  if (has429StatusCode(err)) {
-    return true;
-  }
-  const networkError = err?.networkError;
-  if (Array.isArray(networkError)) {
-    return networkError.some(has429StatusCode);
-  }
-  return false;
-};
-
-const has429StatusCode = (error: any): boolean =>
-  error?.statusCode === 429 || error?.status === 429 || error?.message?.includes('429');
